feat: add POS editing, reviewed flag, and Syriac search to admin UI
- syriac_words.part_of_speech column editable per row
- reviewed flag with Done button — default view shows only unreviewed words
- Show reviewed toggle reveals completed rows with green tint
- Search auto-detects Syriac vs English input
- No POS filter to find the 641 NULL-POS words
- v2_add_pos.sql / v2_detect_pos.py / v2_propagate_pos.sql for POS migration

+++ scripts/v2_search_function.sql
@@ -4,7 +4,7 @@
 
 CREATE OR REPLACE FUNCTION search_dictionary_v2(query_text TEXT, lang TEXT)
 RETURNS TABLE(
-    id            BIGINT,
+    id            INTEGER,
     english       TEXT,
     part_of_speech TEXT,
     variants      JSONB
